Store bucketID for temp objects
