+++ server/pkg/controller/object_cleanup.go
@@ -166,8 +166,10 @@ func (c *ObjectCleanupController) removeUnreportedObjects() int {
 func (c *ObjectCleanupController) removeUnreportedObject(tx *sql.Tx, t ente.TempObject) error {
 	// TODO: object_cleanup
 	// This should use the DC from TempObject (once we start persisting it)
-	// dc := t.DataCenter
-	dc := c.S3Config.GetHotDataCenter()
+	dc := t.BucketId
+	if dc == "" {
+		dc = c.S3Config.GetHotDataCenter()
+	}
 
 	logger := log.WithFields(log.Fields{
 		"task":        "remove-unreported-objects",
@@ -232,7 +234,7 @@ func (c *ObjectCleanupController) addCleanupEntryForObjectKey(objectKey string, 
 	err := c.Repo.AddTempObject(ente.TempObject{
 		ObjectKey:   objectKey,
 		IsMultipart: false,
-		DataCenter:  dc,
+		BucketId:    dc,
 	}, expirationTime)
 	return stacktrace.Propagate(err, "")
 }
@@ -247,7 +249,7 @@ func (c *ObjectCleanupController) AddMultipartTempObjectKey(objectKey string, up
 		ObjectKey:   objectKey,
 		IsMultipart: true,
 		UploadID:    uploadID,
-		DataCenter:  dc,
+		BucketId:    dc,
 	}, expiry)
 	return stacktrace.Propagate(err, "")
 }
